requirements.txt
google-generativeai>=0.3.0
tenacity>=8.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
typing-extensions>=4.0.0
